conditional install of spark

+++ useful/mac-env-install-new.sh
@@ -115,17 +115,31 @@ fi
 
 brew install scala -vd
 
-# TODO if you want the newest spark install with brew
+# TODO
 #brew install apache-spark -vd
 
-echo 'installing spark'
-wget https://archive.apache.org/dist/spark/spark-3.0.3/spark-3.0.3-bin-hadoop3.2.tgz -P ~/Downloads
-mkdir -p ~/hadoop/spark-3.0.3
-tar -xvzf ~/Downloads/spark-3.0.3-bin-hadoop3.2.tgz -C ~/hadoop/spark-3.0.3 --strip 1
-rm -f ~/Downloads/spark-3.0.3-bin-hadoop3.2.tgz
-echo 'export SPARK_HOME=~/hadoop/spark-3.0.3' >> ~/.zshrc
-echo 'export PATH=$SPARK_HOME/bin:$PATH' >> ~/.zshrc
-echo 'installed spark'
+if [[ ! $(command -v spark-submit) == "" ]]; then
+
+    printf "spark exists\n"
+
+    spark-shell --version
+
+    printf 'if you want the newest spark install with brew run:\n\n'
+
+    echo '  brew install apache-spark -vd\n'
+
+else
+
+  echo 'installing spark'
+  wget https://archive.apache.org/dist/spark/spark-3.0.3/spark-3.0.3-bin-hadoop3.2.tgz -P ~/Downloads
+  mkdir -p ~/hadoop/spark-3.0.3
+  tar -xvzf ~/Downloads/spark-3.0.3-bin-hadoop3.2.tgz -C ~/hadoop/spark-3.0.3 --strip 1
+  rm -f ~/Downloads/spark-3.0.3-bin-hadoop3.2.tgz
+  echo 'export SPARK_HOME=~/hadoop/spark-3.0.3' >> ~/.zshrc
+  echo 'export PATH=$SPARK_HOME/bin:$PATH' >> ~/.zshrc
+  echo 'installed spark'
+
+fi
 
 
 brew install kubectl -vd
